refactor: extract browse dialogs mixin to gui/browse_dialogs.py
